usb: k1x_udc_core: support power-supply notify vbus disconnect
USB Device disconnect detection rely on vbus detection, this is done
by using extcon or otg driver based on internal ID/VBUS circuit. But
for some boards, the vbus is connected to a charger, thus no way to
obtain vbus status directly.

Most application works fine without the disconnect since there is
a bus reset after reconnection. But if the f_fs application rely
on the -ECONNRESET on usb request to do some software reinit, it
may cause problem.

To fix problems with this case, we add a notifier connected to power
supply driver similar to extcon.

Change-Id: I58c07dd42419589ef77d0a3b4b2efddd2ab8293a
Signed-off-by: Junzhong Pan <junzhong.pan@spacemit.com>